refactor: decompor tipos de dashboard por dominio
