+++ src/app/routes/dashboard/types/mock-dashboard-data-types.ts
@@ -1,0 +1,4 @@
+export type { Customer } from "@/domains/customers/types/customer-types"
+export type { InventoryItem } from "@/domains/inventory/types/inventory-item-types"
+export type { MenuItem } from "@/shared/types/menu-item-types"
+export type { SalesData } from "./dashboard-analytics-types"
