+++ src/domains/auth/types/user-types.ts
@@ -1,0 +1,17 @@
+import type { UserRole } from "@/shared/types/user-role-types"
+
+export interface User {
+	id: string
+	email: string
+	name: string
+	avatar?: string
+	role: UserRole
+	createdAt: Date
+}
+
+export interface UserRestaurant {
+	organizationId: string
+	restaurantName: string
+	role: UserRole
+	isOwner: boolean
+}
