+++ src/domains/customers/types/customer-types.ts
@@ -1,0 +1,13 @@
+export interface Customer {
+	id: string
+	organizationId: string
+	name: string
+	email: string
+	phone: string
+	totalOrders: number
+	totalSpent: number
+	loyaltyPoints: number
+	lastVisit: Date
+	preferences?: string[]
+	avatar?: string
+}
