+++ src/domains/menu/types/menu-item-types.ts
@@ -1,0 +1,5 @@
+export type {
+	MenuItem,
+	Recipe,
+	RecipeIngredient,
+} from "@/shared/types/menu-item-types"
