+++ src/domains/notifications/types/notification-types.ts
@@ -1,0 +1,9 @@
+export interface Notification {
+	id: string
+	type: "info" | "success" | "warning" | "error"
+	title: string
+	message: string
+	read: boolean
+	createdAt: Date
+	actionUrl?: string
+}
