+++ src/shared/types/commerce-types.ts
@@ -1,0 +1,3 @@
+export type PaymentMethod = "cash" | "credit" | "debit" | "pix" | "meal_voucher"
+
+export type OrderType = "dine_in" | "takeaway" | "delivery"
