+++ src/shared/types/user-role-types.ts
@@ -1,0 +1,1 @@
+export type UserRole = "owner" | "manager" | "cashier" | "kitchen" | "waiter"
